# specs/003-complete-builder/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PointApp Builder API
  version: 1.0.0
  license:
    name: Private and proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Authenticated Builder control-plane and public signed PointApp content-plane contract.
    Browser UI visibility never grants authority; every protected operation is authorized server-side.
servers:
  - url: https://appbuilder.pointatx.org
    description: Production
...
paths:
...
  /api/drafts:
    get:
      tags: [Drafts]
      summary: List drafts
      operationId: listDrafts
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - in: query
          name: state
          schema: { type: string, enum: [active, all], default: active }
      responses:
        '200':
          description: Draft summaries.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Page' }
                  - type: object
                    required: [items]
                    properties:
                      {
                        items:
                          { type: array, items: { $ref: '#/components/schemas/DraftSummary' } },
                      }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Drafts]
      summary: Create or duplicate a draft
      operationId: createDraft
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              additionalProperties: false
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
                duplicateRevisionId: { $ref: '#/components/schemas/Id' }
      responses:
        '201':
          description: Draft and initial revision.
          content: { application/json: { schema: { $ref: '#/components/schemas/DraftDetail' } } }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
  /api/drafts/{draftId}:
    get:
      tags: [Drafts]
      summary: Read a draft and current manifest
      operationId: getDraft
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/DraftId' }]
      responses:
        '200':
          {
            description: Draft with current manifest.,
            content: { application/json: { schema: { $ref: '#/components/schemas/DraftDetail' } } },
          }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Drafts]
      summary: Rename archive or recover a draft
      operationId: updateDraft
      security: [{ sessionCookie: [] }]
      parameters:
        [
          { $ref: '#/components/parameters/DraftId' },
          { $ref: '#/components/parameters/IdempotencyKey' },
        ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedVersion]
              additionalProperties: false
              properties:
                expectedVersion: { type: integer, minimum: 1 }
                name: { type: string, minLength: 1, maxLength: 80 }
                state: { type: string, enum: [active, archived] }
      responses:
        '200':
          {
            description: Updated draft.,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/DraftSummary' } } },
          }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/drafts/{draftId}/revisions:
    get:
      tags: [Drafts]
      summary: List immutable revisions
      operationId: listRevisions
      security: [{ sessionCookie: [] }]
      parameters:
        [
          { $ref: '#/components/parameters/DraftId' },
          { $ref: '#/components/parameters/Cursor' },
          { $ref: '#/components/parameters/Limit' },
        ]
      responses:
        '200':
          description: Immutable revision descriptors.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Page' }
                  - type: object
                    required: [items]
                    properties:
                      {
                        items:
                          { type: array, items: { $ref: '#/components/schemas/RevisionSummary' } },
                      }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Drafts]
      summary: Save an immutable revision
      operationId: saveRevision
      security: [{ sessionCookie: [] }]
      parameters:
        [
          { $ref: '#/components/parameters/DraftId' },
          { $ref: '#/components/parameters/IdempotencyKey' },
        ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expectedParentRevisionId, label, manifest]
              additionalProperties: false
              properties:
                expectedParentRevisionId: { $ref: '#/components/schemas/Id' }
                label: { type: string, minLength: 1, maxLength: 120 }
                manifest: { $ref: '#/components/schemas/Manifest' }
      responses:
        '201':
          {
            description: Immutable saved revision.,
            content:
              { application/json: { schema: { $ref: '#/components/schemas/RevisionSummary' } } },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
  /api/media:
    get:
      tags: [Media]
      summary: List private media assets
      operationId: listMedia
      security: [{ sessionCookie: [] }]
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200':
          description: Private media library page.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Page' }
                  - type: object
                    required: [items]
                    properties:
                      { items: { type: array, items: { $ref: '#/components/schemas/MediaAsset' } } }
        '403': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Media]
      summary: Create a media asset
      operationId: createMedia
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [metadata]
              properties:
                metadata: { $ref: '#/components/schemas/MediaInput' }
                file: { type: string, format: binary, description: Required only for image assets. }
      responses:
        '201':
          {
            description: Ready media asset.,
            content: { application/json: { schema: { $ref: '#/components/schemas/MediaAsset' } } },
          }
        '400': { $ref: '#/components/responses/Problem' }
        '413': { $ref: '#/components/responses/Problem' }
        '415': { $ref: '#/components/responses/Problem' }
  /api/media/{mediaId}:
    get:
      tags: [Media]
      summary: Fetch authenticated media
      operationId: getPrivateMedia
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/MediaId' }]
      responses:
        '200': { description: Image bytes or redirect/metadata response for external media. }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
    patch:
      tags: [Media]
      summary: Archive or recover a media asset
      operationId: updateMediaState
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: '#/components/parameters/MediaId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              additionalProperties: false
              properties: { state: { type: string, enum: [ready, archived] } }
      responses:
        '200':
          description: Updated media lifecycle state.
          content: { application/json: { schema: { $ref: '#/components/schemas/MediaAsset' } } }
        '403': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
...
  /api/audit:
    get:
      tags: [Operations]
      summary: List redacted audit events
      operationId: listAuditEvents
      security: [{ sessionCookie: [] }]
      parameters:
        [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200':
          description: Owner-visible safe audit history.
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/Page' }
                  - type: object
                    required: [items]
                    properties:
                      { items: { type: array, items: { $ref: '#/components/schemas/AuditEvent' } } }
        '403': { $ref: '#/components/responses/Problem' }
  /api/operations:
    get:
      tags: [Operations]
      summary: Read system operations status
      operationId: getOperations
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          {
            description: 'Safe health, release, and capacity overview.',
            content:
              { application/json: { schema: { $ref: '#/components/schemas/OperationsView' } } },
          }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: __Host-pointapp_builder_session }
  parameters:
    Cursor: { in: query, name: cursor, schema: { type: string, maxLength: 256 } }
    Limit:
      { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
    IdempotencyKey:
      { in: header, name: Idempotency-Key, required: true, schema: { type: string, format: uuid } }
    DraftId:
      { in: path, name: draftId, required: true, schema: { $ref: '#/components/schemas/Id' } }
    MediaId:
      { in: path, name: mediaId, required: true, schema: { $ref: '#/components/schemas/Id' } }
...
  responses:
    Problem:
      description: Safe typed error.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  schemas:
    Id: { type: string, format: uuid }
    Role: { type: string, enum: [editor, publisher, administrator, owner] }
...
    DraftSummary:
      type: object
      required: [id, name, state, version, currentRevision, updatedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        name: { type: string }
        state: { type: string, enum: [active, archived] }
        version: { type: integer, minimum: 1 }
        currentRevision: { $ref: '#/components/schemas/RevisionSummary' }
        updatedAt: { type: string, format: date-time }
    DraftDetail:
      allOf:
        - { $ref: '#/components/schemas/DraftSummary' }
        - type: object
          required: [manifest]
          properties: { manifest: { $ref: '#/components/schemas/Manifest' } }
    RevisionSummary:
      type: object
      required: [id, draftId, sequence, checksum, schemaVersion, label, createdBy, createdAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        draftId: { $ref: '#/components/schemas/Id' }
        parentRevisionId: { anyOf: [{ $ref: '#/components/schemas/Id' }, { type: 'null' }] }
        sequence: { type: integer, minimum: 1 }
        checksum: { type: string, pattern: '^[a-f0-9]{64}$' }
        schemaVersion: { type: integer, minimum: 1 }
        label: { type: string }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
    Manifest:
      type: object
      required: [schemaVersion, app, settings, theme, audiences, screens, navigation, campaigns]
      additionalProperties: false
      properties:
        schemaVersion: { type: integer, const: 1 }
        app: { type: object, additionalProperties: true }
        settings: { type: object, additionalProperties: true }
        theme: { type: object, additionalProperties: true }
        audiences:
          { type: array, maxItems: 50, items: { type: object, additionalProperties: true } }
        screens:
          {
            type: array,
            minItems: 1,
            maxItems: 100,
            items: { type: object, additionalProperties: true },
          }
        navigation:
          {
            type: array,
            minItems: 2,
            maxItems: 5,
            items: { type: object, additionalProperties: true },
          }
        campaigns:
          { type: array, maxItems: 100, items: { type: object, additionalProperties: true } }
    MediaInput:
      type: object
      required: [kind, title]
      properties:
        kind: { type: string, enum: [image, audio, video] }
        title: { type: string, minLength: 1, maxLength: 120 }
        altText: { type: string, minLength: 1, maxLength: 500 }
        externalUrl: { type: string, format: uri, pattern: '^https://' }
        captionUrl: { type: string, format: uri, pattern: '^https://' }
        expectedSha256: { type: string, pattern: '^[a-f0-9]{64}$' }
    MediaAsset:
      allOf:
        - { $ref: '#/components/schemas/MediaInput' }
        - type: object
          required: [id, state, mimeType, byteSize, sha256, createdAt]
          properties:
            id: { $ref: '#/components/schemas/Id' }
            state: { type: string, enum: [ready, archived] }
            mimeType: { type: string }
            byteSize: { type: integer, minimum: 0 }
            width: { type: [integer, 'null'], minimum: 1 }
            height: { type: [integer, 'null'], minimum: 1 }
            durationSeconds: { type: [number, 'null'], minimum: 0 }
            sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
            createdAt: { type: string, format: date-time }
...
    ReleaseView:
      type: object
      required:
        [eventId, channel, eventKind, releaseId, revisionId, manifestDigest, keyId, createdAt]
      properties:
        eventId: { $ref: '#/components/schemas/Id' }
        channel: { type: string, enum: [staging, production] }
        eventKind: { type: string, enum: [publish, promote, rollback] }
        releaseId: { $ref: '#/components/schemas/Id' }
        revisionId: { $ref: '#/components/schemas/Id' }
        manifestDigest: { type: string, pattern: '^[a-f0-9]{64}$' }
        keyId: { type: string }
        signature: { type: string, contentEncoding: base64url }
        createdAt: { type: string, format: date-time }
...
    AuditEvent:
      type: object
      required: [id, requestId, action, targetType, outcome, occurredAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        requestId: { $ref: '#/components/schemas/Id' }
        actorGithubUserId: { type: [string, 'null'] }
        actorRole: { anyOf: [{ $ref: '#/components/schemas/Role' }, { type: 'null' }] }
        action: { type: string }
        targetType: { type: string }
        targetId: { type: [string, 'null'] }
        outcome: { type: string, enum: [succeeded, denied, failed] }
        reason: { type: [string, 'null'] }
        occurredAt: { type: string, format: date-time }
    Health:
      type: object
      required: [status, service, environment, checks, time]
      properties:
        status: { type: string, enum: [ok, degraded] }
        service: { type: string, const: pointapp-builder }
        environment: { type: string, enum: [local, production] }
        checks: { type: object, additionalProperties: { type: string, enum: [ok, unavailable] } }
        time: { type: string, format: date-time }
    OperationsView:
      type: object
      required: [health, channels, capacity]
      properties:
        health: { $ref: '#/components/schemas/Health' }
        channels:
          {
            type: object,
            additionalProperties:
              { anyOf: [{ $ref: '#/components/schemas/ReleaseView' }, { type: 'null' }] },
          }
        capacity: { type: object, additionalProperties: { type: integer, minimum: 0 } }
    Page:
      type: object
      required: [items]
      properties: { nextCursor: { type: [string, 'null'] } }
    Problem:
      type: object
      required: [type, title, status, code, requestId]
      additionalProperties: false
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        detail: { type: string }
        requestId: { $ref: '#/components/schemas/Id' }
        fields: { type: object, additionalProperties: { type: array, items: { type: string } } }